[codegen] Validate and write version-dependent fields
During validation, we will now check that all optional fields which are
expected for the current table version are present, and we will write
these correctly.

This wasn't really handled correctly previously, since previously the
only version-dependent fields we were handling were those that were also
nullable, and so were allowed to be absent.