Impl default for generated bitflags
The default will always be empty flags. If this is not acceptable, we
can add an attribute for specifying the desired default value, or an
attribute indicating we won't generate a Default impl.

Default for enums is more annoying: not all enums have a meaningful
default value, and we don't want to guess, so maybe for now the best bet
is to just require the user to implement the trait, or otherwise just
specify an explicit default value for each field in a table that
contains a raw enum.

Another approach altogether (which had mentioned in relation to NameIds)
would be to not use an actual rust enum to represent these enums, and
make them instead by newtype structs with associated constants. This
feels un-rusty, but also an awful lot simpler and closer to representing
the underlying truth.