[font-types] Ensure Tag impls DeserializeOwned
I ran into a bit of a surprise when trying to update fontc to the our
latest release: it was unable to deserialize our Tag type, complaining
that,

> Error("axes[0].tag: invalid type: string \"wght\", expected a borrowed string")

This caused some headscratching, since I couldn't reproduce it in tests.
It turned out to be a funny set of things:

- My implementation involved temporarily deserializing to a &str
- The implementation of Deserialize for &str only works when the input
  can be borrowed
- The deserialization code in fontc uses yaml::from_reader
- from_reader involves reading from a byte stream, which is transient
- ergo you cannot deserialize &str from a reader

The solution is to do a slightly more principled Deserialize
implementation, where I explicitly implement the Visitor trait.