[codegen] Always generate Eq/Ord/Hash where possible
In write-fonts, we will generate these traits for all types.

In read-fonts, the situation is more nuanced. In the case of tables,
derive is not an option, because tables in read-fonts wrap some chunk of
bytes, but that chunk may contain trailing data that is not necessarily
part of the table in question.

More generally, in read-fonts, the semantics of things like 'equality'
are ambiguous. If a table contains an offset to subtable, should
equality be based on the raw value of the offset, or based recursively
on the contents of the child tables?

At some point we may wish to revisit this qusetion, but for now I am
punting. We *will* derive extra traits in read-fonts, but only for
records, and only when they contain only non-offset scalar values.