[read|write-fonts] make offset_to_axis_values nullable
Fixes https://github.com/googlefonts/fontc/issues/582

OTS gives a warning that our offset_to_axis_values is != 0 even though the count of axis_values == 0.
This fixes it

Note that this is a breaking change in that read-fonts offset_to_axis_values now returns an Option<Result<...>> instead of just a Result.