Preserve format of value records after reading
ValueRecords have an explicit format, which is generally stored once in
their parent table. In the case of records with device or variation
tables, we need to know the format in order to write them correctly,
since it's possible for these fields to be legitimately null, and we
cannot distinguish between a record which where a field is None because
it is not included versus when it is intended to be null.

Previously, we would use the format when reading the table but would
then discard it. This led to a bug where we would not know the correct
format when we went to write the table back.

With this patch, we hold onto the format after reading, and pass it to
write-fonts if we convert to write-fonts types. This ensures that we
roundtrip tables correctly in this case.